Refactor: move aspirateur models to repondeur
