+++ aspirateur/setup.py
@@ -8,5 +8,5 @@ setup(
     packages=find_packages(),
     zip_safe=False,
     include_package_data=True,
-    install_requires=["dataclasses"],
+    install_requires=[],
 )
