+++ repondeur/zam_repondeur/fetch/senat/amendements/__init__.py
@@ -7,9 +7,8 @@ from urllib.parse import urljoin
 import requests
 from selectolax.parser import HTMLParser
 
-from zam_aspirateur.amendements.models import Amendement
-
 from zam_repondeur.fetch.exceptions import NotFound
+from zam_repondeur.fetch.models import Amendement
 from zam_repondeur.fetch.senat.senateurs import fetch_and_parse_senateurs, Senateur
 
 from .parse import parse_from_csv, parse_from_json
