+++ repondeur/zam_repondeur/views/lectures.py
@@ -160,7 +160,7 @@ class ListAmendements:
                 Amendement.organe == self.lecture.organe,
             )
             .order_by(
-                case([(Amendement.position.is_(None), 1)], else_=0),
+                case([(Amendement.position.is_(None), 1)], else_=0),  # type: ignore
                 Amendement.position,
                 Amendement.num,
             )
